requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
idna==2.10
lxml==4.6.3
requests==2.25.1
robotframework==4.0
robotframework-pythonlibcore==2.2.1
robotframework-seleniumlibrary==5.1.3
selenium==3.141.0
soupsieve==2.2.1
tqdm==4.59.0
urllib3==1.26.4
webdrivermanager==0.10.0
